fix server to only be available for CLI mode
also fix a bug in cadence adapter where the middleware was not polled
before submitting a metric
