+++ src/cadence.rs
@@ -36,6 +36,7 @@ where
     fn emit(&self, raw_metric: &str) -> io::Result<usize> {
         let cooked_metric = Metric::new(raw_metric.as_bytes().to_vec());
         let mut next = self.next.lock().unwrap();
+        next.poll();
         let result = next.submit(cooked_metric);
 
         match result {
