+++ src/middleware/mod.rs
@@ -1,6 +1,4 @@
-use std::io::ErrorKind;
 use std::net::{ToSocketAddrs, UdpSocket};
-use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
@@ -14,6 +12,9 @@ pub mod allow_tag;
 pub mod cardinality_limit;
 pub mod deny_tag;
 pub mod tag_cardinality_limit;
+
+#[cfg(feature = "cli")]
+pub mod server;
 
 const BUFSIZE: usize = 8192;
 
@@ -140,76 +141,6 @@ impl Middleware for Upstream {
 
     fn poll(&mut self) -> Result<(), Overloaded> {
         self.timed_flush();
-        Ok(())
-    }
-}
-
-pub struct Server<M> {
-    socket: UdpSocket,
-    middleware: M,
-}
-
-impl<M> Server<M>
-where
-    M: Middleware,
-{
-    pub fn new(listen: String, middleware: M) -> Result<Self, Error> {
-        let socket = UdpSocket::bind(listen)?;
-        // An acceptable balance between busyloop and responsiveness to signals.
-        socket.set_read_timeout(Some(Duration::from_secs(1)))?;
-        Ok(Server { socket, middleware })
-    }
-
-    pub fn run(mut self) -> Result<(), Error> {
-        // if sending this large udp dataframes happens to work randomly, we should not be the
-        // one that breaks that setup.
-        let mut buf = [0; 65535];
-
-        let stop = Arc::new(AtomicBool::new(false));
-
-        #[cfg(feature = "cli")]
-        {
-            // This block is basically useless on windows. Would need to implement as a full fledged
-            // service.
-            #[cfg(not(windows))] // No SIGHUP on windows.
-            signal_hook::flag::register(signal_hook::consts::SIGHUP, Arc::clone(&stop))?;
-            signal_hook::flag::register(signal_hook::consts::SIGINT, Arc::clone(&stop))?;
-            signal_hook::flag::register(signal_hook::consts::SIGTERM, Arc::clone(&stop))?;
-        }
-
-        while !stop.load(Ordering::Relaxed) {
-            let (num_bytes, _app_socket) = match self.socket.recv_from(buf.as_mut_slice()) {
-                Err(err) => match err.kind() {
-                    // Different timeout errors might be raised depending on platform.
-                    ErrorKind::WouldBlock | ErrorKind::TimedOut => {
-                        // Allow the middlewares to do any needed bookkeeping.
-                        let _ = self.middleware.poll();
-                        continue;
-                    }
-                    _ => return Err(Error::from(err)),
-                },
-                Ok(s) => s,
-            };
-            for raw in buf[..num_bytes].split(|&x| x == b'\n') {
-                if raw.is_empty() {
-                    continue;
-                }
-
-                let raw = raw.to_owned();
-                let metric = Metric::new(raw);
-
-                let mut carryover_metric = Some(metric);
-                while let Some(metric) = carryover_metric.take() {
-                    if let Err(Overloaded { metric }) = self
-                        .middleware
-                        .poll()
-                        .and_then(|()| self.middleware.submit(metric))
-                    {
-                        carryover_metric = metric;
-                    }
-                }
-            }
-        }
         Ok(())
     }
 }
